requirements.txt
fastapi
uvicorn
pydantic
deepface
face-alignment
dlib
opencv-python
matplotlib
numpy
scikit-image
plotly
requests
Pillow
selenium
pandas
sklearn
scikit-learn
tqdm
numba
torch